frontend: Ignore whitespace inside of braces, parens, and brackets
In meson newlines are only significant as statement separators, but
don't mean anything in places where only an expression is accepted. This
means that inside of `()`, `[]`, and `{}`, we can add newlines freely
and they are not significant.

This patch doesn't appear work correctly with dicts, though I'm not sure
if it's just the `as_string()` method that's broken or if there
something actually wrong.