hir: Start bilding the high level IR
This is going to remain fairly close to the Meson DSL representation. My
thinking right now is that we'll have this at a high level, and we'll
then run optimizations to lower things down closer to the raw arguments,
with backend IRs that take this and get it down into abstractions of
their rules, then write out those rules.

At least, that's the thought for the moment.