mir/passes/value_numbering: introduce
This adds a value numbering pass. All it does is go through and
annotations for which "version" of the variable this is. This doesn't
re-value something that's already been valued. Taht might prove to be
annoying, but I'm going to go with it for now.

If we do need to re-value, we probably need to write a "de-numbering"
pass to run before the value_numbering pass, due to revisiting the same
block twice.

Fixes #29