mir/passes: insert phi nodes
With this we insert phi nodes, which puts us into an SSA from. This is
allows us to know exactly what value we have at any point in the
program, so we know that in cases like:
```meson
if host_machine.system() == 'windows'
  lib = 'foolib'
else
  lib = 'otherlib'
endif
if meson.get_compiler('c').find_library(lib)
```
We know that lib is a phi of either `foolib` or `otherlib`. This is
particularly useful for meson, where most variables are immutable, but
shadowable (there are, unfortunately, some exceptions). With this in
place we can start constant folding, and therefor linking!

Fixes #54