mir: fixup Phis after branch pruning
When we prune a branch, we need to check phis, and make sure that any
phis pointing to a value that is about to be removed is replaced by an
identifier to the value not about to be replaced. Then we can use
constant folding later to do away with any cases of duplicate values.