mir/passes: fix the order we visit branches in
We want to visit false branches before true branches, but that's not
what we're doing. Among other things, this means that we don't have to
track visited branches because the final condition will always have an
empty false branch, this means that we will never visit a node until all
of it's parents have been visited.