mir/passes: Add a value_usage pass
This is the complement to the value_numbering pass. While value
numbering looks at assignments and allows for the insertion of phi
nodes, this pass takes advantage of those numbers and phi nodes to map
usages to assignments. This will allow for constant folding and constant
prop to easily work.