Replace run-tests with meson test targets
Inspired by muon, I've moved all tests into the meson tests. This puts
all the tests together and is generally nice. I'd still like to get rid
of the shell script eventually, but for now this is a nice improvement.
This makes the unittest results produce invalid junit, which is a known
upstream issue for meson.

Fixes: #89