mir/passes: write combined gvn pass
This pass both adds global value numbers, and inserts phis. It's
convenient to do this in a single pass because we can number
assignments and usages at the same time, and the phis will automatically
point to the right place.

This results in some changes to how the variables are numbered, but it
isn't wrong, it's just different.