mir: Add a python script to generate function argument validators from XML
I'm not 100% happy with using Python for this, it would be nice to use a
C++ generator instead, but this was relatively quick and easy to get
going. This doesn't wire the generator in, only implement it, and only
enough of it to handle the one case currently in XML: files.