mir: remove Instruction
This was really not a helpful abstraction, it just made the code very
complicated, and it's only real purpose was to hold the variable
information. For now this has been moved to the Object, but the plan is
to move it out of the Object altogether and use `load` and `store`
types for setting and unsetting values.